feat(tables): uniformiser les tableaux paginés
